+++ tests/testthat/test-validate_identifiers.R
@@ -14,39 +14,39 @@ test_that(".check_required_identifiers works as expected", {
   data.table::setnames(df, unlist(ids))
 
   # Single mapping all exist.
-  obs <- gDRutils:::.check_required_identifiers(df, req_ids = req_ids, id_map = ids)
+  obs <- .check_required_identifiers(df, req_ids = req_ids, id_map = ids)
   expect_equal(obs, NULL)
 
   # Single mapping more ids than required exist.
   excess_map_ids <- ids
   excess_map_ids <- c(excess_map_ids, list(test = "excess"))
-  obs <- gDRutils:::.check_required_identifiers(df, req_ids = req_ids, id_map = excess_map_ids)
+  obs <- .check_required_identifiers(df, req_ids = req_ids, id_map = excess_map_ids)
   expect_equal(obs, NULL)
 
   # Single mapping more columns than required exist.
   fewer_map_ids <- ids
   fewer_req_ids <- req_ids[seq(length(req_ids) - 1)]
   fewer_map_ids <- fewer_map_ids[req_ids]
-  obs <- gDRutils:::.check_required_identifiers(df, req_ids = fewer_req_ids, id_map = fewer_map_ids)
+  obs <- .check_required_identifiers(df, req_ids = fewer_req_ids, id_map = fewer_map_ids)
   expect_equal(obs, NULL)
 
   # Single mapping required id value does not exist.
   single_map_ids <- ids
   single_map_ids[[1]] <- "Cinderella"
-  obs <- gDRutils:::.check_required_identifiers(df, req_ids = req_ids, id_map = single_map_ids)
+  obs <- .check_required_identifiers(df, req_ids = req_ids, id_map = single_map_ids)
   exp <- "specified value identifier(s): 'Cinderella' do not exist for standardized identifier(s): 'duration'\n"
   expect_equal(obs, exp)
 
   # Missing identifier.
   missing_map_ids <- ids
   missing_map_ids[[1]] <- NULL
-  expect_error(gDRutils:::.check_required_identifiers(df, req_ids = req_ids, id_map = missing_map_ids),
+  expect_error(.check_required_identifiers(df, req_ids = req_ids, id_map = missing_map_ids),
     regex = sprintf("required identifiers: '%s' missing in 'id_map'", names(ids)[[1]]))
 
   # Polymapping. 
   poly_map_ids <- ids
   poly_map_ids[[1]] <- c("Cinderella", "Mulan")
-  expect_error(gDRutils:::.check_required_identifiers(df, req_ids = req_ids, id_map = poly_map_ids),
+  expect_error(.check_required_identifiers(df, req_ids = req_ids, id_map = poly_map_ids),
     regex = sprintf("more than one identifier value found for required identifiers: '%s'", names(poly_map_ids)[1]))
 })
 
@@ -63,25 +63,25 @@ test_that(".check_polymapped_identifiers works as expected", {
   data.table::setnames(df, unlist(id_map))
 
   # All singletons.
-  obs <- gDRutils:::.check_polymapped_identifiers(df, exp_one_ids, id_map)
+  obs <- .check_polymapped_identifiers(df, exp_one_ids, id_map)
   expect_equal(obs, NULL)
 
   # Some polymappings.
   some_poly_map <- id_map
   some_poly_map[[1]] <- c(some_poly_map[[1]], "extra_item") 
-  obs <- gDRutils:::.check_polymapped_identifiers(df, exp_one_ids, id_map = some_poly_map)
+  obs <- .check_polymapped_identifiers(df, exp_one_ids, id_map = some_poly_map)
   expect_equal(obs, "more than one mapping for identifier(s): 'duration'\n")
 
   # All polymappings.
   all_poly_map <- id_map
   all_poly_map <- lapply(seq_along(all_poly_map), function(x) c(all_poly_map[[x]], "extra_item"))
   names(all_poly_map) <- names(id_map)
-  obs <- gDRutils:::.check_polymapped_identifiers(df, exp_one_ids, id_map = all_poly_map)
+  obs <- .check_polymapped_identifiers(df, exp_one_ids, id_map = all_poly_map)
   exp <- paste0(names(id_map), collapse = ", ")
   expect_equal(obs, sprintf("more than one mapping for identifier(s): '%s'\n", exp))
 })
 
-test_that("gDRutils:::.modify_polymapped_identifiers works as expected", {
+test_that(".modify_polymapped_identifiers works as expected", {
   # Set up.
   nrow <- 5
   exp_one_ids <- get_expect_one_identifiers()
@@ -94,26 +94,26 @@ test_that("gDRutils:::.modify_polymapped_identifiers works as expected", {
   data.table::setnames(df, unlist(id_map))
 
   # All singletons.
-  obs <- gDRutils:::.modify_polymapped_identifiers(df, exp_one_ids, id_map)
+  obs <- .modify_polymapped_identifiers(df, exp_one_ids, id_map)
   expect_equal(obs, id_map)
 
   # Some polymappings.
   some_poly_map <- id_map
   some_poly_map[[1]] <- c(some_poly_map[[1]], "extra_item") 
-  obs <- gDRutils:::.modify_polymapped_identifiers(df, exp_one_ids, id_map = some_poly_map)
+  obs <- .modify_polymapped_identifiers(df, exp_one_ids, id_map = some_poly_map)
   expect_equal(obs, id_map)
 
   # All polymappings.
   all_poly_map <- id_map
   all_poly_map <- lapply(seq_along(all_poly_map), function(x) c(all_poly_map[[x]], "extra_item"))
   names(all_poly_map) <- names(id_map)
-  obs <- gDRutils:::.modify_polymapped_identifiers(df, exp_one_ids, id_map = all_poly_map)
+  obs <- .modify_polymapped_identifiers(df, exp_one_ids, id_map = all_poly_map)
   expect_equal(obs, id_map)
 
   # Multiple valid polymappings.
   multi_valid_map <- id_map
   multi_valid_map[[1]] <- c(multi_valid_map[[1]], multi_valid_map[[2]])
-  expect_error(gDRutils:::.modify_polymapped_identifiers(df, exp_one_ids, id_map = multi_valid_map),
+  expect_error(.modify_polymapped_identifiers(df, exp_one_ids, id_map = multi_valid_map),
     regex = sprintf("multiple valid identifier values found for identifier: '%s'", names(multi_valid_map)[1]))
 })