+++ NEWS.md
@@ -1,3 +1,6 @@
+## gDRutils 1.1.10 - 2024-03-07
+- clean up the package
+
 ## gDRutils 1.1.9 - 2024-03-07
 * simplify keywords